requirements.txt
sarvamai>=0.1.0
python-dotenv>=1.0.0
